Add files via upload
The art of technology from reservation to ordering and storing customer records is the concept of Smart restaurant. An Android application helps the customer to view the fully automated menu ordering system. The Web application also helps the customer to place the order. The ordering process is carried out by scanning the QR code on the reservation table and is used to handle the delivery activity. The order is rooted to the kitchen via a central server. The kitchen having a direct interface displays the order and services based on their priority (first come first serve). Ultrasonic sensor is used for human detection. When humans are detected, the door is opened using servo motors. The food delivery is carried out through the conveyor belt to the customers. The conveyor belt is run using DC motor. The payment is done by scanning the QR code and payment option is also provided in the web application. Feedback option is also provided during the exit of the customers.